Fixed a really weird issue.
For some reason, the controller bindings don't work correctly if
IControllerBinding is an "interface." It seems like the reason is due to
the fact that each class doesn't inherit their own versions of the axio
maps and whatnot. So what this leads to is constant shuffling of the
values and what they are bound to. I don't think we should have it be an
interface or maybe I should come up with another design?