Started work on the map data visualizer.
It currently doesn't work, but the set up for actually making it happen
is there.

There is quite a lot of things I need to do, but for now let's just
focus on the getting it to display pixels into a window and make it
accurate to the data stored in the file.