Got a very basic "blobbing" feature working.
It should probably have something more, but right now all it does is
expand a single pixel into a "circle" (if you can even call it that).

It should help with visualizing certain paths, but I probably need to
turn off the color adding for it to work well visually.

Anyways, that's it.